+++ api/src/main/java/org/yapp/domain/auth/application/oauth/service/OauthService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 import org.yapp.domain.auth.application.jwt.JwtUtil;
 import org.yapp.domain.auth.application.oauth.OauthProvider;
 import org.yapp.domain.auth.application.oauth.OauthProviderResolver;
-import org.yapp.domain.auth.dto.request.OauthLoginRequest;
-import org.yapp.domain.auth.dto.response.OauthLoginResponse;
+import org.yapp.domain.auth.presentation.dto.request.OauthLoginRequest;
+import org.yapp.domain.auth.presentation.dto.response.OauthLoginResponse;
 import org.yapp.domain.user.User;
 import org.yapp.domain.user.dao.UserRepository;
