+++ api/src/main/java/org/yapp/domain/auth/presentation/dto/response/OauthLoginResponse.java
@@ -1,4 +1,4 @@
-package org.yapp.domain.auth.dto.response;
+package org.yapp.domain.auth.presentation.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Getter
 public class OauthLoginResponse {
-    private boolean registered;
-    private String accessToken;
-    private String refreshToken;
+  private boolean registered;
+  private String accessToken;
+  private String refreshToken;
 }
