+++ api/src/main/java/org/yapp/domain/auth/presentation/dto/response/OauthUserInfoResponse.java
@@ -1,4 +1,4 @@
-package org.yapp.domain.auth.dto.response;
+package org.yapp.domain.auth.presentation.dto.response;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -6,5 +6,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 public class OauthUserInfoResponse {
-    private String id;
+  private String id;
 }
